// frontend/rodia-monorepo/apps/mobile/src/shared/api/generated/schemas/counterOfferAcceptResponse.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface CounterOfferAcceptResponse {
  counterOfferId?: number;
  quoteId?: number;
  matchId?: number;
  driverId?: number;
  counterOfferStatus?: string;
  quoteStatus?: string;
  matchStatus?: string;
  paymentRequired?: boolean;
  nextAction?: string;
}